+++ libp2p-test/.eslintrc.js
@@ -1,0 +1,7 @@
+module.exports = {
+  parserOptions: {
+    ecmaVersion: 'latest',
+    sourceType: 'module',
+  },
+  rules: {},
+}
